Fix positional tuple fields in Display impl (#20)
* Fix positional tuple fields in Display impl

The field names (numeric indices) were not being stripped for tuple
variants. This caused errors when attempting to output Display impls
which did not reference every sequential index.

This change makes it possible to reference arbitrary tuple fields
as intended.

* Remove an unnecessary unwrap

`fmt::Write` for `String` is infallible. We can safely ignore the
`Result`.